#2916 comment: moved logger handler definitions and Base boto3 client definition to common.utilities to avoid circular imports
