--- api/common/log_color_handler.py
@@ -1,108 +1,0 @@
-#
-#   Thiscovery API - THIS Institute’s citizen science platform
-#   Copyright (C) 2019 THIS Institute
-#
-#   This program is free software: you can redistribute it and/or modify
-#   it under the terms of the GNU Affero General Public License as
-#   published by the Free Software Foundation, either version 3 of the
-#   License, or (at your option) any later version.
-#
-#   This program is distributed in the hope that it will be useful,
-#   but WITHOUT ANY WARRANTY; without even the implied warranty of
-#   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#   GNU Affero General Public License for more details.
-#
-#   A copy of the GNU Affero General Public License is available in the
-#   docs folder of this project.  It is also available www.gnu.org/licenses/
-#
-
-# Adapted from https://xsnippet.org/359377/
-# with modifications described in https://stackoverflow.com/a/55724913
-
-import epsagon
-import sys
-import logging
-
-
-class _AnsiColorizer(object):
-    """
-    A colorizer is an object that loosely wraps around a stream, allowing
-    callers to write text to the stream in a particular color.
-
-    Colorizer classes must implement C{supported()} and C{write(text, color)}.
-    """
-    _colors = dict(black=30, red=31, green=32, yellow=33,
-                   blue=34, magenta=35, cyan=36, white=37)
-
-    def __init__(self, stream):
-        self.stream = stream
-
-    @classmethod
-    def supported(cls, stream=sys.stdout):
-        """
-        A class method that returns True if the current platform supports
-        coloring terminal output using this method. Returns False otherwise.
-        """
-        if not stream.isatty():
-            return False  # auto color only on TTYs
-        try:
-            import curses
-        except ImportError:
-            return False
-        else:
-            try:
-                try:
-                    return curses.tigetnum("colors") > 2
-                except curses.error:
-                    curses.setupterm()
-                    return curses.tigetnum("colors") > 2
-            except:
-                raise
-                # guess false in case of error
-                return False
-
-    def write(self, text, color):
-        """
-        Write the given text to the stream in the given color.
-
-        @param text: Text to be written to the stream.
-
-        @param color: A string label for a color. e.g. 'red', 'white'.
-        """
-        color = self._colors[color]
-        self.stream.write('\x1b[%s;1m%s\x1b[0m' % (color, text))
-
-
-class ColorHandler(logging.StreamHandler):
-
-    def __init__(self, stream=sys.stderr):
-        super(ColorHandler, self).__init__(_AnsiColorizer(stream))
-
-    def emit(self, record):
-        msg_colors = {
-            logging.DEBUG: "green",
-            logging.INFO: "blue",
-            logging.WARNING: "yellow",
-            logging.ERROR: "red"
-        }
-
-        color = msg_colors.get(record.levelno, "blue")
-        self.stream.write(self.format(record) + "\n", color)
-
-
-class EpsagonHandler(logging.Handler):
-    ssm_client = None
-    running_tests = None
-
-    def emit(self, exception_instance):
-        if self.ssm_client is None:
-            from common.ssm_utilities import Ssm  # delayed import to circumvent circular import problem
-            ssm_client = Ssm()
-            self.running_tests = ssm_client.get_parameter('running-tests', prefix='/thiscovery/')
-
-        if self.running_tests == 'false':
-            epsagon.error(exception_instance)
-        elif self.running_tests == 'true':
-            pass
-        else:
-            raise AttributeError(f'AWS SSM parameter /thiscovery/running-tests is neither "true" nor "false": {self.running_tests}')
